### 2023-03-03 #5103 @csoneson
👋🏻 @souzamonteiro - I will handle your submission and start by recruiting at least two suitable reviewers. If you have suggestions, please feel free to mention them here!

In the meanwhile, a couple of comments that you can already address:
- Your paper is quite a bit longer than the JOSS guidelines (see [here](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain)). A typical JOSS paper is 250-1000 words (yours is currently 3105), and it should contain a 'Statement of need' section. Could you please adapt your paper to be more in line with the JOSS guidelines?
- Also please take a look at the invalid DOIs flagged by @editorialbot above - mostly the issue seems to be that the DOIs include the `'https://doi.org/'` prefix.
