### 2023-03-20 #5103 @csoneson
👋🏻 @souzamonteiro - just wanted to check in here - did you have a chance to look into shortening the paper to adhere closer to the JOSS guidelines?

> In the meanwhile, a couple of comments that you can already address:
> Your paper is quite a bit longer than the JOSS guidelines (see [here](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain)). A typical JOSS paper is 250-1000 words (yours is currently 3105), and it should contain a 'Statement of need' section. Could you please adapt your paper to be more in line with the JOSS guidelines?
> Also please take a look at the invalid DOIs flagged by @editorialbot above - mostly the issue seems to be that the DOIs include the 'https://doi.org/' prefix.
