### 2023-03-20 93a772a
Reduce the length of the article to meet the JOSS editor's request
