+++ paper/paper.bib
@@ -46,7 +46,7 @@
 	number = {4},
 	pages = {1234-1248},
 	year  = {2016},
-	doi = {https://doi.org/10.21577/1984-6835.20160088},
+	doi = {10.21577/1984-6835.20160088},
 	URL = {http://static.sites.sbq.org.br/rvq.sbq.org.br/pdf/v8n4a16.pdf}}
 
 @article{Casteigts2012,
@@ -58,8 +58,8 @@
 	pages = {387-408},
 	year  = {2012},
 	publisher = {Taylor \& Francis},
-	doi = {https://doi.org/10.1080/17445760.2012.668546},
-	URL = {https://doi.org/10.1080/17445760.2012.668546}}
+	doi = {10.1080/17445760.2012.668546},
+	URL = {10.1080/17445760.2012.668546}}
 
 @article{Monteiro2014,
 	title = {Evolution based on chromosome affinity from a network perspective},
@@ -68,7 +68,7 @@
 	pages = {276-283},
 	year = {2014},
 	issn = {0378-4371},
-	doi = {https://doi.org/10.1016/j.physa.2014.02.019},
+	doi = {10.1016/j.physa.2014.02.019},
 	url = {https://www.sciencedirect.com/science/article/pii/S0378437114001204},
 	author = {Roberto Luiz S. Monteiro and José Roberto A. Fontoura and Tereza K. G. Carneiro and Marcelo A. Moret and Hernane B. B. Pereira}}
 
@@ -79,7 +79,7 @@
 	pages = {121542},
 	year = {2019},
 	issn = {0378-4371},
-	doi = {https://doi.org/10.1016/j.physa.2019.121542},
+	doi = {10.1016/j.physa.2019.121542},
 	url = {https://www.sciencedirect.com/science/article/pii/S0378437119309069},
 	author = {Eder Johnson de Area Leão Pereira and Paulo Jorge Silveira Ferreira and Marcus Fernandes da Silva and José Garcia Vivas Miranda and Hernane B. B. Pereira}}
 
@@ -90,7 +90,7 @@
 	pages = {194-208},
 	year = {2019},
 	issn = {0951-8320},
-	doi = {https://doi.org/10.1016/j.ress.2019.02.013},
+	doi = {10.1016/j.ress.2019.02.013},
 	url = {https://www.sciencedirect.com/science/article/pii/S0951832018308135},
 	author = {Ying Zhou and Chenshuang Li and Lieyun Ding and Przemyslaw Sekula and Peter E. D. Love and Cheng Zhou}}
 
@@ -102,7 +102,7 @@
 	pages = {2236–2249},
 	year = {2021},
 	issn = {2236-2249},
-	doi = {https://doi.org/10.1002/hbm.25362},
+	doi = {10.1002/hbm.25362},
 	url = {https://pubmed.ncbi.nlm.nih.gov/33570215/},
 	author = {Dan Zhu and Tengfei Yuan and Junfeng Gao and Qiang Xu and Kaizhong Xue and Wenshuang Zhu and Jie Tang and Feng Liu and Junping Wang and Chunshui Yu}}
 
@@ -113,7 +113,7 @@
 	pages = {101025},
 	year = {2022},
 	issn = {0911-6044},
-	doi = {https://doi.org/10.1016/j.jneuroling.2021.101025},
+	doi = {10.1016/j.jneuroling.2021.101025},
 	url = {https://www.sciencedirect.com/science/article/pii/S0911604421000415},
[…]
