### 2023-03-20 e8383f7
Reduce the length of the article to meet the JOSS editor's request
