### 2023-03-31 #5103 @adavidzh
> 👋🏻 @adavidzh, @tomalrussell - would you be interested in reviewing this submission for JOSS?
>
> > CNATool - Complex Network Analysis Tool

Thank you @csoneson. I am afraid that I am not able to commit to review this contribution at this time.
