### 2023-04-01 #5103 @csoneson
@paulgovan - would you be interested in reviewing this submission for JOSS?

> CNATool - Complex Network Analysis Tool
