+++ paper/paper.bib
@@ -1,188 +1,189 @@
 @article{Newman2003,
-  author = {M. E. J. Newman},
-  title = {The structure and function of complex networks},
-  journal = {SIAM Review},
-  volume = {45},
-  number = {2},
-  year = {2003},
-  pages = {167--256}
+  author={M. E. J. Newman},
+  title={The structure and function of complex networks},
+  journal={SIAM Review},
+  volume={45},
+  number={2},
+  year={2003},
+  pages={167--256}
 }
 
 
 @article{Pereira2007,
-	author = {Hernane Borges de Barros Pereira and Mario Cezar Freitas and Renelson Ribeiro Sampaio},
-	title = {Fluxos de informação e conhecimentos para inovações no arranjo produtivo local de confecções em Salvador, Bahia},
-	journal = {DataGramaZero -- Revista de Ciência da Informação},
-	volume = {8},
-	number = {4},
-	year = {2007}}
+	author={Hernane Borges de Barros Pereira and Mario Cezar Freitas and Renelson Ribeiro Sampaio},
+	title={Fluxos de informação e conhecimentos para inovações no arranjo produtivo local de confecções em Salvador, Bahia},
+	journal={DataGramaZero -- Revista de Ciência da Informação},
+	volume={8},
+	number={4},
+	year={2007}}
 
 @article{Moody2005,
-	ISSN = {00029602, 15375390},
-	URL = {http://www.jstor.org/stable/10.1086/421509},
-	author = {James Moody and Daniel McFarland and Skye Bender‐deMoll},
-	journal = {American Journal of Sociology},
-	number = {4},
-	pages = {1206--1241},
-	publisher = {The University of Chicago Press},
-	title = {Dynamic Network Visualization},
-	urldate = {2022-12-29},
-	volume = {110},
-	year = {2005}}
+	ISSN={00029602, 15375390},
+	URL={http://www.jstor.org/stable/10.1086/421509},
+	author={James Moody and Daniel McFarland and Skye Bender‐deMoll},
+	journal={American Journal of Sociology},
+	number={4},
+	pages={1206--1241},
+	publisher={The University of Chicago Press},
+	title={Dynamic Network Visualization},
+	urldate={2022-12-29},
+	volume={110},
+	year={2005}}
 
 @article{Prieto2008,
-	author = {Beatriz Prieto and Fernando Tricas and Juan J. Merelo and Antonio Mora and Alberto Prieto},
-	title = {Visualizing the evolution of a web-based social network},
-	journal = {Journal of Network and Computer Applications},
-	number = {31},
-	year = {2008},
-	pages = {677--698}}
+	author={Beatriz Prieto and Fernando Tricas and Juan J. Merelo and Antonio Mora and Alberto Prieto},
+	title={Visualizing the evolution of a web-based social network},
+	journal={Journal of Network and Computer Applications},
+	number={31},
+	year={2006},
+	pages={677--698},
+	doi={10.1016/j.jnca.2007.11.003}}
 
 @article{Vieira2016,
-	author = {Rodrigo P. Vieira and Roberto Luiz S. Monteiro and Hernane B. B. Pereira and Jailson B. de Andrade and Lilian L. N. Guarieiro},
-	title = {Redes de Colaboração Científica do INCT de Energia e Ambiente},
-	journal = {Revista Virtual Química},
-	volume = {8},
-	number = {4},
-	pages = {1234-1248},
-	year  = {2016},
-	doi = {10.21577/1984-6835.20160088},
-	URL = {http://static.sites.sbq.org.br/rvq.sbq.org.br/pdf/v8n4a16.pdf}}
[…]
