### 2023-04-03 #5103 @paulgovan
Hi @csoneson, I’m familiar with network analysis but not Maiascript, so I’m probably not the best reviewer for this.
