### 2023-04-04 #5103 @souzamonteiro
> @souzamonteiro - thanks! Please note that JOSS papers are also required to have a `Statement of need` section. (also it looks like some DOIs are still missing or invalid).

Thanks for the feedback. I already did what was asked.
