### 2023-04-15 #5373 @editorialbot
[REVIEW]: CNATool - Complex Network Analysis Tool
**Submitting author:** @souzamonteiro (<a href="http://orcid.org/0000-0002-3931-5953">Roberto L. S. Monteiro</a>)
**Repository:** https://github.com/souzamonteiro/cnatool
**Branch with paper.md** (empty if default branch):
**Version:** 2.1.1
