### 2023-04-19 ccb7446
Updated DOI for The New Science of Networks paper
