+++ paper/paper.bib
@@ -201,7 +201,7 @@
 @article{Watts2004,
 	ISSN={03600572, 15452115},
 	URL={http://www.jstor.org/stable/29737693}, 
-	doi={10.2307/29737693},
+	doi={10.1146/annurev.soc.30.020404.104342},
 	author={Duncan J. Watts},
 	journal={Annual Review of Sociology},
 	pages={243--270},
