### 2023-04-20 #5373 @souzamonteiro
> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
> OK DOIs
>
> - 10.1016/j.jnca.2007.11.003 is OK
> - 10.21577/1984-6835.20160088 is OK
> - 10.1080/17445760.2012.668546 is OK
> - 10.1016/j.physa.2014.02.019 is OK
> - 10.1016/j.physa.2019.121542 is OK
> - 10.1016/j.ress.2019.02.013 is OK
> - 10.1002/hbm.25362 is OK
> - 10.1016/j.jneuroling.2021.101025 is OK
> - 10.2224/sbp.7681 is OK
> - 10.1016/j.physa.2020.125012 is OK
> - 10.1016/j.physa.2021.125848 is OK
> - 10.1016/j.socnet.2021.06.001 is OK
> - 10.1038/nrn2575 is OK
> - 10.1007/s00429-014-0836-3 is OK
> - 10.1098/rstb.2000.0551 is OK
> - 10.1016/j.physa.2015.08.024 is OK
> - 10.1016/S0079-6123(07)68012-1 is OK
> - 10.7155/jgaa.00108 is OK
> - 10.1145/321992.321993 is OK
> - 10.1007/978-3-540-77891-2_6 is OK
> - 10.1103/PhysRevLett.87.198701 is OK
> - 10.1016/0378-8733(78)90021-7 is OK
> - 10.1016/0378-8733(79)90002-9 is OK
> - 10.1007/978-3-642-21934-4_4 is OK
> - 10.2307/3033543 is OK
> - 10.1038/30918 is OK
> - 10.1016/j.neucom.2005.12.070 is OK
> - 10.1016/S0378-4371(02)00736-7 is OK
> - 10.1016/S0378-4371(02)01545-5 is OK
>
> MISSING DOIs
>
> - None
>
> INVALID DOIs
>
> - 10.2307/29737693 is INVALID

Corrigido
