### 2023-04-29 4f3e11d
Adjustments regarding the software license were made to meet the requests of the JOSS reviewer
