### 2023-05-15 souzamonteiro/cnatool#2 @tomalrussell
cnatool command-line import errors
Hi @souzamonteiro

I've successfully installed `cnatool` as a command-line tool. This works both from a local checkout of the git repository, and from `npm`.

When running `cnatool`, I see several `require` errors:

```
$ cnatool
Cannot find module 'algebrite'
Require stack:
- /usr/lib/node_modules/cnatool/bin/cnatool.js
Cannot find module 'gpu.js'
Require stack:
- /usr/lib/node_modules/cnatool/bin/cnatool.js
Cannot find module 'web-worker'
Require stack:
- /usr/lib/node_modules/cnatool/bin/cnatool.js
Cannot find module 'cross-blob'
Require stack:
- /usr/lib/node_modules/cnatool/bin/cnatool.js
CNATool Command Line Interface (CLI)
Usage: cnatool [options] [network.net] [--] [arguments]
```

Should these packages somehow be specified as requirements?

Are there example commands that would fail without these modules?
