### 2023-05-15 souzamonteiro/cnatool#3 @tomalrussell
Built files and dependencies added to git
This is a general code/repository review comment.

I would consider removing all of the following files from git:
- files that can be specified as dependencies (`node_modules`, `sigma.js-1.2.1.zip`, `tmp`, some of `js`)
- files that can be built from the source code ( some of `js`, `docs`, `build`, `bin`)
- duplicate copies (`manual/*` is copied and versioned under `docs/`)
- files that aren't otherwise used here (`fonts/maya-allstars-font.zip`, `images/{cloud,liza}*`)
- test outputs (`test.txt`, `test/test.html`?)
- unused submodules (`.gitmodules` has https://github.com/a-nsilva/joss-CNATool twice, but the content has moved into this repository with more recent changes)

If possible, also avoid the duplication of built files - do we need all of `src`, `js` and `bin`?

Could `maiascript` and `libmaia` be specified as dependencies and installed or required?

This would make it easier for a reviewer or new contributor to orient themselves and focus on the most important files in the repository.
