### 2023-05-18 #5373 @souzamonteiro
Hi:

I added the following text to the READMe.md file:

I The program depends on the following packages:
- algebrite
Install using the command:
`npm install -g algebrite`
- web-worker
Install using the command:
`npm install -g web-worker`
- cross-blob
Install using the command:
`npm install -g cross-blob`
- gpu.js
Install using the command:
`npm install -g gpu.js`

Regarding the choice of the MaiaScript language, I am the author of the language and CNATOOL is one of the demonstration programs for the applications of the language. That's why the libraries used in CNATOOL were written in MaiaScript.

About the mentioned:

API documentation linked from the online tool to CNATool API Reference and CNA API Reference both return 404 not found. Some docs are available as HTML files within the git repository.

The site frequently undergoes updates and sometimes some features move around and need to be updated. I will provide the necessary corrections.

Regarding the other recommendations: I am arranging to make changes to the text and other additions and modifications.
