### 2023-05-22 souzamonteiro/cnatool#2 @ati-ozgur
I am also having the same errors.
Please fix the package.json (https://github.com/souzamonteiro/cnatool/blob/main/package.json) so that dependencies will be automatically installed with "npm install -g cnatool" command
