### 2023-05-22 #5373 @ati-ozgur
I think adding dependencies to readme makes this software hard to use. I added a comment to [issue ](https://github.com/souzamonteiro/cnatool/issues/2) asking adding dependencies to package.json. Please do so.
