### 2023-05-22 souzamonteiro/cnatool#4 @ati-ozgur
I would like this also either in readme or in documentation. This usage should be linked to if it is not in the readme.
