+++ CONTRIBUTORS.md
@@ -1,0 +1,3 @@
+## CONTRIBUTORS
+
+To contribute to this project, please clone the repository on github, make the changes, and send the link to your repository containing your contributions to support@maiascript.com.
