### 2023-06-29 souzamonteiro/cnatool#2 @tomalrussell
`npm i -g cnatool` gives the same errors as before - an updated version should be pushed to npmjs.com/package/cnatool

`npm i` from a checkout of this repository at 8e6df7bf00acda3ab258f9cd8041cf4240ec7467 required several C/C++ libraries installing, all of which should be documented (and more, if I overlooked any that I happened to have already on this machine) - for me, on Ubuntu with a bit of guesswork from `pkg-config` error messages:
```
sudo apt install libx11-dev libxi-dev libgl1mesa-dev
```
Presumably these are required by a particular npm package - is it that `gpu.js` depends on `gl` depends on these libraries? ([gpu.js docs](https://www.npmjs.com/package/gpu.js/v/2.4.0#installation))
