### 2023-07-06 souzamonteiro/cnatool#2 @souzamonteiro
> `npm i -g cnatool` gives the same errors as before - an updated version should be pushed to npmjs.com/package/cnatool
>
> `npm i` from a checkout of this repository at [8e6df7b](https://github.com/souzamonteiro/cnatool/commit/8e6df7bf00acda3ab258f9cd8041cf4240ec7467) required several C/C++ libraries installing, all of which should be documented (and more, if I overlooked any that I happened to have already on this machine) - for me, on Ubuntu with a bit of guesswork from `pkg-config` error messages:
>
> ```
> sudo apt install libx11-dev libxi-dev libgl1mesa-dev
> ```
>
> Presumably these are required by a particular npm package - is it that `gpu.js` depends on `gl` depends on these libraries? ([gpu.js docs](https://www.npmjs.com/package/gpu.js/v/2.4.0#installation))

I updated the package on NPM with the fixes. There are actually no C/C++ library dependencies that need to be installed manually.  Some Linux distributions do not include some libraries that are dependencies on Node.js packages and not on CNATOOL itself. I have tested the installation without any issues on Ubuntu 20.04 and Ubuntu 22.04. Either using "npm install cnatool", or entering the directory downloaded from github and typing "npm install .". Libraries that are missing from your system are used by OpenGL Dev System. The Mesa3D dev packages are currently only installed on some distributions if you install video card specific drivers. Can be it...
