### 2023-07-06 souzamonteiro/cnatool#3 @souzamonteiro
Hello:
I removed the node_modules directory and copies of the manual from the docs directory. The manual folder exists to meet the needs of other magazines that ask for this folder so I will have to leave the manuals in it. The sigma.js-1.2.1.zip file will remain because I use this particular version of the package and don't want users to use another one that I can't guarantee will work. The other directories I find useful as well as the test reports because I often need these results on github for several reasons. The js, src etc directories are used in the online version and not all users are able to recreate them, so I decided to keep them as well. Thanks for the comments. maiascript and libmaia are not dependencies that can be installed automatically as CNATOOL uses these specific versions of the libraries and not those available in the npm repository. They are there to ensure that graduate students using the tool can run it directly when downloading from github.
