### 2023-07-06 souzamonteiro/cnatool#4 @souzamonteiro
> How should a new user find or make use of the [example files](https://github.com/souzamonteiro/cnatool/tree/main/examples)?
>
> I'd suggest adding a section to the documentation - how to load examples into the web tool, or how to run them through the command-line tool, and then what to expect or look for in the tool/output. What are the examples meant to illustrate?
>
> What are the sources of these datasets? Are any of them synthetic or created just for this project? I'd suggest adding citations and references as appropriate.

The article originally had 3000 words, where all the information that has been requested was duly cited, but it was requested that the article be reduced to 100 words, which greatly impaired the quality of the work.
