### 2023-07-06 souzamonteiro/cnatool#4 @tomalrussell
Could some of those 3000 words be added to the README or documentation? That would still be useful for users of the tool, even if it's not in scope for the paper.
