### 2023-07-06 souzamonteiro/cnatool#2 @tomalrussell
Can confirm that the install from npmjs.com works for me with v2.0.6 :+1:

I would recommend linking through to those `gpu.js` installation instructions, `npm i` was failing for me without those header libraries (suggested as `mesa-common-dev libxi-dev`) available.. in case anyone else has the same setup issue.

(Out of the original scope of this issue, but worth mentioning: ideally, the basic software setup could be verified in a CI action - have you considered running the build script and test suite in a GitHub action or similar automated system?)
