### 2023-07-06 souzamonteiro/cnatool#3 @tomalrussell
Hi @souzamonteiro, thanks for taking a look at this, and explaining the need for some things to be included here.

Great, keeping the `manuals` directory makes sense.

It looks like the deletion of [`node_modules`](https://github.com/souzamonteiro/cnatool/tree/b23764e93acb7d7d55fe5e2a144fe6242404a279/node_modules) still needs to be pushed to `main`

For info, `sigma.js` v1.2.1 can be installed reliably with `npm install sigma@1.2.1 --save-dev`, with minified files built and saved for you in `node_modules/sigma/build/`. You could for example consider removing `tmp` and `sigma.js-1.2.1.zip` and `js/sigma.js` and replacing them with a copy of those minified files - that would let you shorten the long list of `<script...` tags in `index.html` too, and would retain the benefits of having the library commited and versioned in this repository.

Okay, I acknowledge the need for `maiascript` and `libmaia` to be vendored and included to pin to specific versions.

Do you still need those files that aren't referenced anywhere else (fonts/maya-allstars-font.zip, images/{cloud,liza}*) or the unused submodules (.gitmodules)?
