### 2023-07-06 souzamonteiro/cnatool#2 @tomalrussell
Just running `build.sh` and `test.sh` - all seem to build and pass, but `maiascript` now requires `websql`:

    Cannot find module 'websql'

I ran `npm i --save websql` to add the package, which fixes the warning. I'd recommend updating `package.json` in the repository.
