### 2023-07-12 souzamonteiro/cnatool#2 @souzamonteiro
> Just running `build.sh` and `test.sh` - all seem to build and pass, but `maiascript` now requires `websql`:
>
> ```
> Cannot find module 'websql'
> ```
>
> I ran `npm i --save websql` to add the package, which fixes the warning. I'd recommend updating `package.json` in the repository.

Hi:
I added the dependency in package.json. But none of the dependencies that were previously pointed out, in addition to this one, prevent the program from working. All are optional and were not placed in package.json to prevent installation on systems where they are not supported. For example, websql and gpu.js will not install on some Linux distributions and some versions of node.js. I can't maintain support for all Linux distributions and versions of node.js. Nor is it the scope of this project. The program is only tested on Ubuntu 20.04 and 22.04 and just download from github in zip format for Windows. I don't recommend installing via npm on Windows because the dependencies that were included may not be supported. However, the HTML GUI will work on all platforms, including Android and iOS. But the program must be downloaded directly from github.
