### 2023-07-12 souzamonteiro/cnatool#3 @souzamonteiro
> Hi @souzamonteiro, thanks for taking a look at this, and explaining the need for some things to be included here.
>
> Great, keeping the `manuals` directory makes sense.
>
> It looks like the deletion of [`node_modules`](https://github.com/souzamonteiro/cnatool/tree/b23764e93acb7d7d55fe5e2a144fe6242404a279/node_modules) still needs to be pushed to `main`
>
> For info, `sigma.js` v1.2.1 can be installed reliably with `npm install sigma@1.2.1 --save-dev`, with minified files built and saved for you in `node_modules/sigma/build/`. You could for example consider removing `tmp` and `sigma.js-1.2.1.zip` and `js/sigma.js` and replacing them with a copy of those minified files - that would let you shorten the long list of `<script...` tags in `index.html` too, and would retain the benefits of having the library commited and versioned in this repository.
>
> Okay, I acknowledge the need for `maiascript` and `libmaia` to be vendored and included to pin to specific versions.
>
> Do you still need those files that aren't referenced anywhere else (fonts/maya-allstars-font.zip, images/{cloud,liza}*) or the unused submodules (.gitmodules)?

sygma.js is a GUI dependency and not a command line tool. The graphical interface runs directly in the browser and does not need to be installed. The sygma is provided as it is so that users who want to can run the program simply by downloading and unzipping it from github and double clicking on the index.html file. It would not be useful to place it as a dependency and would even reduce the ease of installation. Most users are from the sociology area and do not understand anything about programming and would not know how to perform the necessary steps. For this reason I choose to keep the sygma.js library inside the package, as it is now. Thanks for the suggestion.
