### 2023-07-12 souzamonteiro/cnatool#4 @souzamonteiro
> How should a new user find or make use of the [example files](https://github.com/souzamonteiro/cnatool/tree/main/examples)?
>
> I'd suggest adding a section to the documentation - how to load examples into the web tool, or how to run them through the command-line tool, and then what to expect or look for in the tool/output. What are the examples meant to illustrate?
>
> What are the sources of these datasets? Are any of them synthetic or created just for this project? I'd suggest adding citations and references as appropriate.

USING THE SAMPLE FILES

They were created by the author and are available under the same license as the program. I noticed that there were some files created by students that were exclusive to their master's projects. I removed them and put a pointer about the example files in README.md.
