### 2023-07-12 6048c4a
Added dependencies, removed unnecessary files, improved documentation and made some changes to the paper
