+++ paper/paper.md
@@ -47,15 +47,34 @@ bibliography: paper.bib
 
 # Summary
 
-CNATool is an online program for analyzing complex and social networks. It was developed to allow quick and simplified analysis of network graphs from any device connected to the Internet. It supports graphs in Pajek and JSON formats, creates graphs of artificial networks, displays graphs, selects the layout algorithm and displays its properties (average degree, density, average clustering coefficient, average shortest path, diameter and efficiency). It also displays detailed properties of the vertices and saves the graph in Pajek, JSON and SVG formats, in addition to generating HTML reports.
+CNATool is an innovative online program designed for analyzing complex and social networks. It provides researchers with a convenient and streamlined approach to network graph analysis, accessible from any Internet-connected device. With support for Pajek and JSON formats, CNATool offers a range of features, including the creation of artificial networks, customizable layout algorithms, and detailed visualization of network properties. Researchers can explore essential network properties such as average degree, density, clustering coefficient, shortest path, diameter, and efficiency. CNATool also provides in-depth analysis of vertex properties, enables graph saving in multiple formats (Pajek, JSON, SVG), and generates comprehensive HTML reports. The tool caters to researchers involved in social and biological network analysis, offering a user-friendly platform for exploring network properties and uncovering patterns. Potential applications include studying social networks, analyzing time-varying networks, investigating functional architecture, examining memory deterioration, exploring the impact of centrality, analyzing accident risks, modeling opinion diffusion, and studying collaborative networks among criminal members. CNATool empowers researchers to delve deeper into the intricacies of complex networks, facilitating meaningful insights and discoveries.
 
 # Introduction
 
-The analysis of complex networks has become frequent, considering that many social, biological and technological phenomena can be represented as network diagrams (also known as Graphs), allowing not only a visual analysis of the problem, but the use of tools developed for a given application involving graph theory [@Newman2003], in solving problems different from those originally thought. @Pereira2007, in their study on LPAs (local productive arrangements), used networks theory to study the processes of cooperation and collaboration between companies that make up an LPA of confections by analyzing the data using the Pajek software [@Moody2005; @Prieto2008]. @Vieira2016 studied collaboration processes between researchers at an important research institute, using similar techniques applying Gephi software in their analysis, also used by @Casteigts2012 in their research on time-varying networks (TVGs). @Monteiro2014 proposed a model appropriate to explain evolution of species in an affinity network, based on an evolutive algorithm that incorporates the same network properties. @Pereira2019, analyzed the stock exchanges for 20 countries in a multiscale network identifying the linkages between markets. @Zhou2019 used complex network theory to monitor coupled risks. @Zhu2021 applied centrality calculations to study the functional architecture of healthy individuals trying to determine their regulatory mechanism. @Chang2021 studied the deterioration of associative memory in individuals with mild amnesia, analyzing semantic clusters. @Li2019 analyzed the effect of centrality of directors on social network in charitable donations. @Guo2020 showed how the Chinese construction industry applies complex network analysis to study accident risks. @Zhang2021 proposed a model of public opinion diffusion in social networks applying the theory of scale-free networks. @Krajewski2022 studied the effect of centrality in collaborative networks among criminal members of the Italian-American mafia.
[…]
