+++ CONTRIBUTORS.md
@@ -1,3 +1,7 @@
 ## CONTRIBUTORS
 
+Contributions are always appreciated. You can contribute with code improvements, documentation improvements, implementation of new features or testing programs and sending us a report with the results you get.
+
 To contribute to this project, please clone the repository on github, make the changes, and send the link to your repository containing your contributions to support@maiascript.com.
+
+Please follow JSDOC's self-documentation precepts for documenting code entered and add test scripts where appropriate.
