+++ examples/dlfNAT001.txt
@@ -1,2569 +1,0 @@
-promise	N
-threat	N
-knowledgebased	NOTFOUND
-economy	N
-{S}	
-unitedstates	NOTFOUND
-university	N
-create	V
-bridge	N
-physics	N
-biology	N
-{S}	
-france	N+PR
-question	N
-israeli	N+Hum
-research	N
-deal	N+Conc
-{S}	
-easterneurope	NOTFOUND
-still	N+Conc
-strap	V
-cash	N
-{S}	
-hungarian	N+Hum
-biology	N
-i	N
-world	N
-class	N
-{S}	
-japanese	N
-budget	N
-boost	N
-science	N
-funding	N
-{S}	
-software	N
-problem	N
-delay	N
-unitedstates	NOTFOUND
-spacecraft	N
-meeting	N
-asteroid	N
-{S}	
-taking	N
-knowledge	N
-poor	A
-{S}	
-worldbank	NOTFOUND
-invest	V
-global	A
-science	N
-base	N
-{S}	
-unitedstates	NOTFOUND
-spirit	N+Abst
-i	N
-willing	A
-funds	N
-are	N+Unit
-still	N+Conc
-weak	A
-{S}	
-science	N
-move	N
-europe	N+PR
-aid	N
-agendum	N
-{S}	
-technology	N
-tradition	N
-clash	N
-india	N+PR
-{S}	
-internet	N
-may	N+Hum
-help	N
-bridge	N
-gap	N
-{S}	
-aids	N
-decision	N
-threaten	V
-drug	N+Conc
-trial	N
-{S}	
-money	N
-talk	N
-loud	A
-research	N
-quality	N
-{S}	
-lysenko	NOTFOUND
-take	V
-dismal	A
-view	N
-malthus	NOTFOUND
-{S}	
-beauty	N
-bart	NOTFOUND
-simpson	NOTFOUND
-effect	N
-{S}	
-english	N
-spanish	N
-science	N
-evaluation	N
-{S}	
-eighteenninetynine	NOTFOUND
-{S}	
-evolutionary	A
-biology	N
-mother	N+Hum
-father	N+Hum
-surprise	N
-geneticagreement	NOTFOUND
-{S}	
-biophysics	N
-may	N+Ntime
-force	N
-be	V
-{S}	
-shrike	N+Anl
-mobility	N
-{S}	
-geochemistry	N
-colloidal	A
-culprit	N+Hum
-contamination	N
-{S}	
-plant	N+Conc
-virus	N+Anl
-new	ADV
-tool	N+Conc
-swiss	N+Hum
-army	N+HumColl
-knife	N+Conc
-{S}	
-coral	N
-reef	N
-recruitment	N
-space	N
-time	N
-{S}	
-photosynthesis	N
-making	N
-light	N+Conc
-dark	N
-situation	N
-{S}	
-timing	N
-end	N
-nocturnal	A
-sleep	N
-{S}	
-air	N
-traffic	N
-may	N+Hum
-increase	N
-cirrus	N
-cloudiness	N
-{S}	
-high	N
-metabolic	A
-rate	N
-running	N
-bird	N+Anl
-{S}	
-thomas	NOTFOUND
-jefferson	N+PR+Hum
-paternity	N
-case	N
-{S}	
-starry	A
-night	N
-{S}	
-wizard	N+Hum
-menlo	NOTFOUND
-park	N
-{S}	
-hypnotism	N
-rise	N
-anaesthesia	N
-{S}	
-mother	N+Hum
-love	N
-selective	A
-advantage	N
-{S}	
-brown	N
-dwarf	N+Hum
-star	N+Hum
-fail	V
-{S}	
-giant	N+Hum
-periodic	A
-flare	N
-soft	N
-ray	N
[…]
