### 2023-07-12 #5373 @souzamonteiro
> Hi, thanks for the changes, these go some way to addressing my comments, but I would like to see more in response some points. I've updated the [review comment](https://github.com/openjournals/joss-reviews/issues/5373#issuecomment-1527569315) above, restating briefly here:
>
> Software issues
>
> * Installation instructions [cnatool command-line import errors souzamonteiro/cnatool#2](https://github.com/souzamonteiro/cnatool/issues/2)
> * Version-control best practices [Built files and dependencies added to git souzamonteiro/cnatool#3](https://github.com/souzamonteiro/cnatool/issues/3)
> * Sources for example data [How to use examples souzamonteiro/cnatool#4](https://github.com/souzamonteiro/cnatool/issues/4)
>
> Documentation
>
> * Statement of need in docs: the software provides a library, a CLI tool and a web-based interface. How do each of these serve different needs? The README and User Guide would be improved with a high-level overview and introduction.
> * User Guide PDF has been updated in the [GitHub repository](https://github.com/souzamonteiro/cnatool/blob/8e6df7bf00acda3ab258f9cd8041cf4240ec7467/manual/CNATool_en_US.pdf); it should also be updated in the [online tool](https://www.maiascript.com/cnatool/docs/CNATool_en_US.pdf).
> * API docs: links now work from the online tool, but should be updated to latest 2.0.5
> * Community guidelines: A `CONTRIBUTORS.md` has been created, but could be more welcoming and needs extending to cover how to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support.
>
> Paper
>
> * State of the field and references: consider discussing other packages in the field and referencing key dependencies.
> * Summary and quality of writing: needs some restructuring, so the summary is more high-level and the introduction is shorter, allowing more discussion of the software.

Hello:

Thanks for the contributions. We try to respond to all requests: we review the paper; we improved README.md and CONTRIBUTORS.md; we include dependencies in package.json; we tested it on some Linux distributions on Windows and macOS; we removed some sample files that had been entered by students and that we could not confirm the origin; added a paragraph in README.md about the examples; corrected all documentation references and updated the website.
