### 2023-07-12 #5373 @ati-ozgur
I do not believe that all of my issues are resolved. Issues are closed without resolving them.

- test.cmd file deleted instead of updating in repository. This means that this code is unix only now.
- documentation of examples are not updated. I expect an example usage in the readme file.

Only, install from command line is fixed in my opinion.
