+++ CONTRIBUTORS.md
@@ -5,3 +5,5 @@ Contributions are always appreciated. You can contribute with code improvements,
 To contribute to this project, please clone the repository on github, make the changes, and send the link to your repository containing your contributions to support@maiascript.com.
 
 Please follow JSDOC's self-documentation precepts for documenting code entered and add test scripts where appropriate.
+
+Please, any testing or development must be carried out in a UNIX environment and if you want to do this in Windows, you must install Cygwin or MSYS. These tools provide a complete UNIX environment on Windows and can be obtained from [https://www.cygwin.com] and [https://www.msys2.org] respectively. Explanations on how to install and use the programs provided with them can be found on their websites.
