### 2023-07-12 #5373 @souzamonteiro
> I do not believe that all of my issues are resolved. Issues are closed without resolving them.
>
> * test.cmd file deleted instead of updating in repository. This means that this code is unix only now.
> * documentation of examples are not updated. I expect an example usage in the readme file.
>
> Only, install from command line is fixed in my opinion.

I'm sorry, but I believe there is a mistake here. test.cmd is in the directory next to test.sh. But I really don't see the need for it because we want the tool to be developed only in a UNIX environment. This does not exclude Windows. As you know, the Cygwin w MSYS tools provide a complete UNIX environment on top of the Windows operating system. I use these tools in my tests on Windows with no problems. Anyway, the test.cmd file is in the directory as you requested. I am also adding information in README.md and CONTRIBUTORS.md explaining that any testing or development must be performed in a UNIX environment and if the user wants to do this in Windows, Cygwin or MSYS must be installed.

As for README.md, it has been changed and information about example files has been added. I'm now uploading the changes again, because I believe there was a problem with the PUSH.
