### 2023-07-14 #5373 @tomalrussell
Thanks @souzamonteiro for the changes, these have dealt with most of my outstanding points.. I think I'd push for another look at the paper itself.

The paper has a rich set of references to network analysis papers; I still think it would be improved by referencing key dependencies and some related software.

- Please cite [sigma.js](https://www.sigmajs.org/) as a dependency of the web-based tool which provides core graph layout and rendering functionality
- [cytoscape](https://cytoscape.org/) is another actively used visualisation and analysis program in a similar space, with both Java and JavaScript implementations - the JS project may be an interesting contrast with `cnatool` as it can also run in web browser and node JS environments.
- Commonly used packages for general network analysis with a scripting interface include `networkx` (Python) and `igraph` (R, Python, Mathematica, C/C++) - these or other packages could be briefly referenced in the section where batch/command-line processing is discussed.
- The paper claims that the software works with "the most common file formats". Common file formats not supported by `cnatool` include GEXF and GraphML, which may be worth mentioning as a (minor) limitation, for example in the feature comparison table.
