### 2023-07-14 souzamonteiro/cnatool#3 @tomalrussell
Okay, it does make sense in that case to keep a copy of the GUI script dependencies. There would still be advantages in terms of number and size of files, if you included the minified built files, rather than a full copy of the library in both `./tmp` and `./js/sigma.js`. Saving `sigma@1.2.1` as a development dependency (which is what `--save-dev` does) would allow developers of the tool to install (and eventually update) the library automatically, without forcing it on users. Reducing the number of files should make it easier for non-expert users to find the most relevant files, and not be confused by unnecessary or unused files.

I'd still recommend removing:
- files that aren't referenced anywhere else (fonts/maya-allstars-font.zip, images/{cloud,liza}*)
- the unused submodules (.gitmodules)
- the [node_modules](https://github.com/souzamonteiro/cnatool/tree/b23764e93acb7d7d55fe5e2a144fe6242404a279/node_modules) directory
