### 2023-07-14 e7027cf
Removed unnecessary files from sigma.js library and .gitmodules file
