### 2023-07-14 #5373 @souzamonteiro
> Thanks @souzamonteiro for the changes, these have dealt with most of my outstanding points.. I think I'd push for another look at the paper itself.
>
> The paper has a rich set of references to network analysis papers; I still think it would be improved by referencing key dependencies and some related software.
>
> * Please cite [sigma.js](https://www.sigmajs.org/) as a dependency of the web-based tool which provides core graph layout and rendering functionality
> * [cytoscape](https://cytoscape.org/) is another actively used visualisation and analysis program in a similar space, with both Java and JavaScript implementations - the JS project may be an interesting contrast with `cnatool` as it can also run in web browser and node JS environments.
> * Commonly used packages for general network analysis with a scripting interface include `networkx` (Python) and `igraph` (R, Python, Mathematica, C/C++) - these or other packages could be briefly referenced in the section where batch/command-line processing is discussed.
> * The paper claims that the software works with "the most common file formats". Common file formats not supported by `cnatool` include GEXF and GraphML, which may be worth mentioning as a (minor) limitation, for example in the feature comparison table.

Thank you for the considerations. I modified the document as follows:
1) Adding the item "Graph Visualization:</strong> Visualization of graphs: CNATool allows the visualization of graphs, from the graphical interface, thanks to the [@Sigma2023] library, including resources such as dynamic adjustment of the layout of the graph , either through algorithms such as ForceAtlas2, or manually dragging and positioning the vertices" in the section "CNATool: Innovative Network Analysis Program"
2) Adding a column about Cytoscape in the table and the text "An important observation is that the Cytoscape program, although it does not allow its execution directly on the Web, as an application, it allows, through the [@Cytoscape2023] library, the creation of Web programs for visualization and analysis of network graphs. Other popular and powerful software such as [@Igraph2023] and [@NetworkX2023] were not compared as they are not user applications but programming libraries." at the end of the "Statement of Need" session and;
3) I created the "Limitations" section with the text "As a minor limitation, CNATool does not support the GEXF and GraphML formats, however, the tools that support these formats are generally able to export graphs in Pajek (.NET) format, allowing the interoperability of CNATool with these software."
