+++ paper/paper.bib
@@ -429,3 +429,31 @@ doi={10.1007/978-3-642-21934-4_4}}
 	pages={333--347},
 	year={2010},
 	publisher={World Scientific}}
+
+@software{Sigma2023,
+	author = {Sciences-Po médialab and OuestWare},
+	title = {sigma.js},
+	url = {https://www.sigmajs.org},
+	version = {1.2.1},
+	date = {2023-07-14}}
+
+@software{Cytoscape2023,
+	author = {National Institute of General Medical Sciences (NIGMS)},
+	title = {Cytoscape},
+	url = {https://cytoscape.org},
+	version = {3.10.0},
+	date = {2023-07-14}}
+
+@software{Igraph2023,
+	author = {The igraph core team},
+	title = {Igraph},
+	url = {https://igraph.org},
+	version = {0.10.6},
+	date = {2023-07-14}}
+
+@software{NetworkX2023,
+	author = {NetworkX developers},
+	title = {NetworkX},
+	url = {https://networkx.org},
+	version = {3.1},
+	date = {2023-07-14}}
