### 2023-07-14 3fa53f0
Fix in paper table and reference
