### 2023-07-14 c6b3cb2
Fix in paper references
