### 2023-07-26 #5373 @souzamonteiro
>
The DOI:
 10.1007/978-1-4939-7131-2_100771

Is not about NetworkX. It is a site, a software, not a book. There is no DOI for this publication.
