### 2023-07-29 88dffde
Paper updated to meet publisher requests
