+++ paper/paper.bib
@@ -1,12 +1,12 @@
 @article{Newman2003,
-  author={M. E. J. Newman},
-  title={The structure and function of complex networks},
-  journal={SIAM Review},
-  volume={45},
-  number={2},
-  year={2003},
-  pages={167--256}
-}
+	author={M. E. J. Newman},
+	title={The structure and function of complex networks},
+	journal={SIAM Review},
+	volume={45},
+	number={2},
+	year={2003},
+	pages={167--256},
+	doi={10.1137/S003614450342480}}
 
 
 @article{Pereira2007,
@@ -223,7 +223,7 @@
 	author={Hernane B. B. Pereira and I. S. Fadigas and Roberto L. S. Monteiro and Aloísio J. A. Cordeiro and Marcelo A. Moret}}
 
 @incollection{Chatterjee2007,
-	title={Understanding the mind of a worm: hierarchical network structure underlying nervous system function in C. elegans},
+	title={Understanding the mind of a worm: hierarchical network structure underlying nervous system function in \NoCaseChange{C. elegans}},
 	editor={Rahul Banerjee and Bikas K. Chakrabarti},
 	series={Progress in Brain Research},
 	publisher={Elsevier},
@@ -259,7 +259,7 @@
 	issn={0004-5411},
 	url={10.1145/321992.321993},
 	doi={10.1145/321992.321993},
-	journal={J. ACM},
+	journal={Journal of the ACM},
 	month={jan},
 	pages={1–13},
 	numpages={13}}
@@ -277,7 +277,7 @@
 @article{Latora2001,
 	title={Efficient Behavior of Small-World Networks},
 	author={Latora, Vito and Marchiori, Massimo},
-	journal={Phys. Rev. Lett.},
+	journal={Physical Review Letters},
 	volume={87},
 	issue={19},
 	pages={198701},
@@ -320,7 +320,7 @@ publisher={Springer Berlin Heidelberg},
 address={Berlin, Heidelberg},
 pages={537--546},
 isbn={978-3-642-21934-4},
-doi={10.1007/978-3-642-21934-4_4}}
+doi={10.1007/978-3-642-21934-4_44}}
 
 @article{Freeman1977,
 	title={A set of measures of centrality based on betweenness},
@@ -339,7 +339,8 @@ doi={10.1007/978-3-642-21934-4_4}}
 	number={2},
 	pages={163--177},
 	year={2001},
-	publisher={Taylor \& Francis}}
+	publisher={Taylor \& Francis},
+	doi={10.1080/0022250X.2001.9990249}}
 
 @article{Barthelemy2004,
 	title={Betweenness centrality in large complex networks},
@@ -349,7 +350,8 @@ doi={10.1007/978-3-642-21934-4_4}}
 	number={2},
 	pages={163--168},
 	year={2004},
-	publisher={Springer}}
+	publisher={Springer},
+	doi={10.1140/epjb/e2004-00111-4}}
 
 @article{Curado2022,
 	title={Anew centrality measure in dense networks based on two-way random walk betweenness},
@@ -358,7 +360,8 @@ doi={10.1007/978-3-642-21934-4_4}}
 	volume={412},
 	pages={126560},
 	year={2022},
-	publisher={Elsevier}}
+	publisher={Elsevier},
+	doi={10.1016/j.amc.2021.126560}}
 
 @article{Watts1998,
 	title={Collective dynamics of small-world networks},
@@ -379,7 +382,8 @@ doi={10.1007/978-3-642-21934-4_4}}
 	number={3-4},
 	pages={539--546},
 	year={2000},
-	publisher={Elsevier}}
+	publisher={Elsevier},
+	doi={10.1016/S0378-4371(00)00311-3}}
 
 @article{Emmert2006,
[…]
