+++ paper/paper.md
@@ -62,7 +62,7 @@ These examples demonstrate the wide range of applications and tools developed fo
 CNATool is an innovative online program designed specifically for analyzing complex and social networks. With its user-friendly interface and accessibility from any internet-connected device, CNATool provides researchers with a convenient and streamlined approach to network graph analysis.
 Key Features:
 <ul>
-<li><strong>Graph Visualization:</strong> Visualization of graphs: CNATool allows the visualization of graphs, from the graphical interface, thanks to the sigma.js [@Sigma2023] library, including resources such as dynamic adjustment of the layout of the graph, either through algorithms such as ForceAtlas2, or manually dragging and positioning the vertices.
+<li><strong>Graph Visualization:</strong> CNATool allows the visualization of graphs, from the graphical interface, thanks to the sigma.js [@Sigma2023] library, including resources such as dynamic adjustment of the layout of the graph, either through algorithms such as ForceAtlas2, or manually dragging and positioning the vertices.
 <li><strong>Graph File Formats:</strong> CNATool supports Pajek and JSON formats, allowing researchers to import network graphs effortlessly.</li>
 <li><strong>Artificial Networks:</strong> Researchers can create artificial networks within CNATool, enabling customizable experimental analysis.</li>
 <li><strong>Customizable Layout Algorithms:</strong> CNATool offers customizable layout algorithms, facilitating the visualization of network structures.</li>
@@ -71,19 +71,13 @@ Key Features:
 <li><strong>Export and Reports:</strong> Researchers can save graphs in multiple formats, including Pajek, JSON, and SVG, and generate comprehensive HTML reports for documentation purposes.</li>
 </ul>
 
-## Potential 
-
-Applications CNATool can be utilized in various research applications, including the study of social networks, analysis of time-varying networks, exploration of functional architecture in healthy individuals, investigation of associative memory deterioration, examination of the impact of centrality on social networks, analysis of accident risks in the construction industry, modeling of public opinion diffusion, and exploration of collaborative networks among criminal members.
-
-By offering a comprehensive range of features and analysis capabilities, CNATool empowers researchers to delve deeper into the intricacies of complex and social networks, facilitating meaningful insights and discoveries.
-
 # Statement of Need 
 
 @Newman2003 presents the main concepts involved in complex and social network analysis. The author discusses the types of networks, topologies, local and global properties. Regarding the types of networks, Newman highlights social, informational, technological and biological networks. These networks, despite having different natures, present common properties such as number of vertices, number of edges, density [@Pereira2016; @Chatterjee2007], average degree, average clustering coefficient [@Schank2005], average shortest path [@Johnson1977], diameter [@Razzaque2008] and efficiency [@Latora2001]. And even at micro scale, similar parameters are observed, highlighting the clustering coefficient and the closeness [@Freeman1978; @Freeman1979; @Bhardwaj2011] and betweenness [@Freeman1977; @Brandes2001; @Barthelemy2004; @Curado2022] centralities. With regard to topologies, networks of apparently different natures, such as social and biological, often present phenomena common to small-world [@Watts1998; @Marchiori2000; @Emmert2006; @Bakshy2011] and scale-free networks [@Barabasi2002; @Crucitti2003].
 
[…]

[…]
