+++ paper/paper.bib
@@ -223,7 +223,7 @@
 	author={Hernane B. B. Pereira and I. S. Fadigas and Roberto L. S. Monteiro and Aloísio J. A. Cordeiro and Marcelo A. Moret}}
 
 @incollection{Chatterjee2007,
-	title={Understanding the mind of a worm: hierarchical network structure underlying nervous system function in \NoCaseChange{C. elegans}},
+	title={Understanding the mind of a worm: hierarchical network structure underlying nervous system function in  \NoCaseChange{C. elegans}},
 	editor={Rahul Banerjee and Bikas K. Chakrabarti},
 	series={Progress in Brain Research},
 	publisher={Elsevier},
