### 2023-08-01 #5373 @souzamonteiro
> Thanks - no need to tick the boxes, I guess it's a permission issue. The text looks all good, but I have to admit I'm still a bit puzzled by the colors in Figure 1:
>
> > In Figure 1, what do the colors ('low/medium/high') mean for the Platform? I.e., why is Windows yellow for Pajek, but green for SocNetV?
>
> It's not clear to me what 'High' means for a 'Platform' - does it map to the number of platforms supported, or the ease of use or accessibility of these platforms, and is 'High' or 'Low' corresponding to 'good' here? If you want to use colors, I would suggest that e.g. green always corresponds to a 'desirable' state (rather than 'High', since high accuracy is good, but high complexity is not. I find Table 1 easier to read in this respect, since all the rows correspond to desirable properties, so a 'Yes' is always good. In the figure, the 'Complexity' box in itself also has different colors for the different tools, and it's not clear to me if this represents some information.
>
> Once this is fixed I think we're ready to move to the next steps, I'll make another post with a checklist for that.

Hello:

I have altered the figure in order to make it more consistent in terms of colors and also its content. I also modified the caption of the figure to:

"Mind map of concepts involving analysis of complex networks. For calculation accuracy and interface complexity, green means desirable feature, yellow means acceptable feature, and red means undesirable feature."

I hope I have heeded your recommendations.
