### 2023-08-01 #5373 @csoneson
Thank you, I think this is much clearer!

I think we can move on to the next steps - could you take a look at the author tasks in the [post-review checklist](https://github.com/openjournals/joss-reviews/issues/5373#issuecomment-1657066593)? Let me know if anything is unclear.
