### 2023-08-01 #5373 @souzamonteiro
>
[x] Double check authors and affiliations (including ORCIDs)
[x] Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.
I created version 2.1.o, aka Nautilus and publised it on github.
[ ] Archive the release on Zenodo/figshare/etc and post the DOI here.
I do not know how to do this. Can you help me?
[x] Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.
I checked the authors and saw no errors.
[ ] Make sure that the license listed for the archive is the same as the software license.
I'm not sure I understand: the license for the program is Apache 2. Do I need to change this or put it somewhere else?
