### 2023-08-01 #5373 @csoneson
Great - to archive your repository on Zenodo, see instructions here: https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content.

Regarding
> [ ] Make sure that the license listed for the archive is the same as the software license.
I'm not sure I understand: the license for the program is Apache 2. Do I need to change this or put it somewhere else?

This refers to the Zenodo archive (once it is created) - make sure that the same license is indicated there.
