### 2023-08-01 #5373 @souzamonteiro
> That seems correct. Could you please edit the metadata of the Zenodo record to match the title and author list of the JOSS paper?

That's it: I made the changes to the Zenodo registry and added Andréia Rita da Silva's ORCID, as it was missing.
