### 2023-08-02 #5373 @csoneson
Just copying this here for reference:

> The DOI 10.1007/978-1-4939-7131-2_100771 is not about NetworkX. It is a site, a software, not a book. There is no DOI for this publication.
